refactor(config-ui): move the base layout to routes/layout
