+++ config-ui/src/routes/layout/tips-transition.css
